requirements.txt
Flask==0.10.1
cloudant==2.3.1
python-dotenv
json
